typescript on frontend almost finished/working in browser

+++ public/src/TaskController.ts
@@ -17,10 +17,12 @@ export class TaskController {
 
   init = async () => {
     this.model.localData = await this.model.dataService.get();
-    this.view = new UIView(this.model.localData[0]);
-    this.view.render(this.root, 'beforeend');
-    this.setData();
-    this.addEventListeners();
+    if (this.model.localData) {
+      this.view = new UIView(this.model.localData[0]);
+      this.view.render(this.root, 'beforeend');
+      this.setData();
+      this.addEventListeners();
+    }
   };
 
   setData = () => {
@@ -30,83 +32,96 @@ export class TaskController {
   };
 
   setCategories = () => {
-    for (let i = 0; i < this.model.localData[1].length; i++) {
-      this.model.localData[1][i] = new CategoryVO(
-        this.model.localData[1][i].index,
-        this.model.localData[1][i].label,
-        this.model.localData[1][i].color,
-        this.model.localData[1][i]._id
-      );
-      const picker = <HTMLInputElement>(
-        document.getElementById(this.model.localData[1][i].index)
-      );
-      picker.value = this.model.localData[1][i].color;
+    if (this.model.localData) {
+      for (let i = 0; i < this.model.localData[1].length; i++) {
+        this.model.localData[1][i] = new CategoryVO(
+          this.model.localData[1][i].index,
+          this.model.localData[1][i].label,
+          this.model.localData[1][i].color,
+          this.model.localData[1][i]._id
+        );
+        const picker = <HTMLInputElement>(
+          document.getElementById(this.model.localData[1][i].index.toString())
+        );
+        picker.value = this.model.localData[1][i].color;
+      }
+      this.colorAllTasks();
     }
-    this.colorAllTasks();
   };
 
   setColumns = () => {
-    for (let i = 0; i < this.model.localData[0].length; i++) {
-      this.model.localData[0][i] = new ListVO(
-        this.model.localData[0][i].label,
-        this.model.localData[0][i].items,
-        this.model.localData[0][i]._id
-      );
+    if (this.model.localData) {
+      for (let i = 0; i < this.model.localData[0].length; i++) {
+        this.model.localData[0][i] = new ListVO(
+          this.model.localData[0][i].label,
+          this.model.localData[0][i].items,
+          this.model.localData[0][i]._id
+        );
+      }
     }
   };
 
   setTasks = () => {
-    this.model.localData[0].forEach(column => {
-      for (let i = 0; i < column.items.length; i++) {
-        column.items[i] = new TaskVO(
-          column.items[i].index,
-          column.items[i].body,
-          column.items[i].category,
-          column.label,
-          column.items[i]._id,
-          column.items[i].columnID
-        );
-      }
-    });
-  };
-
-  colorTask = task => {
-    // not sure if this is a good idea
-    const picker = <HTMLInputElement>document.getElementById(task.category);
-    document.getElementById(task.index)!.style.backgroundColor = picker.value;
-  };
-
-  colorAllTasks = () => {
-    this.model.localData[0].forEach(column => {
-      column.items.forEach(task => {
-        const picker = <HTMLInputElement>document.getElementById(task.category);
-        document.getElementById(task.index)!.style.backgroundColor =
-          picker.value;
+    if (this.model.localData) {
+      this.model.localData[0]?.forEach(column => {
+        for (let i = 0; i < column.items.length; i++) {
+          column.items[i] = new TaskVO(
+            column.items[i].index,
+            column.items[i].body,
+            column.items[i].category,
+            column.label,
+            column.items[i]._id,
+            column.items[i].columnID
+          );
+        }
       });
-    });
+    }
+  };
+
+  colorTask = (task: TaskVO) => {
+    // not sure if this is a good idea
+    // task.index stringified
+    const picker = <HTMLInputElement>document.getElementById(task.category);
+    document.getElementById(task.index.toString())!.style.backgroundColor =
+      picker.value;
+  };
+
+  // task.index stringified
+  colorAllTasks = () => {
+    if (this.model.localData) {
+      this.model.localData[0]?.forEach(column => {
+        column.items.forEach((task: TaskVO) => {
+          const picker = <HTMLInputElement>(
+            document.getElementById(task.category)
+          );
+          document.getElementById(
+            task.index.toString()
+          )!.style.backgroundColor = picker.value;
+        });
+      });
+    }
   };
 
   createTask = () => {
     const body = <HTMLTextAreaElement>document.querySelector('#new-card-text');
-    const category = this.checkCategory();
+    const category: string = this.checkCategory();
     const column = document.querySelectorAll('.col-body')[1];
-    // I put here 2 nulls to have 6 arguments matching 6 parameters
-    const task = taskFactory(
-      this.model.localData[2].counter++,
+    const task: TaskVO = taskFactory(
+      this.model.localData![2].counter++,
       body.value,
       category,
       // not sure here, to do - maybe I don't need column property at all
       column.label
     );
-    this.model.localData[0][0].items.push(task);
+    this.model.localData![0][0].items.push(task);
     body.value = '';
     return task;
   };
 
-  checkCategory = () => {
+  checkCategory = (): string => {
     const radioButtons = document.getElementsByName('importance');
     let importance;
-    let category;
+    let category: string = '';
     radioButtons.forEach(btn => {
       const button = btn as HTMLInputElement;
       if (button.checked) {
@@ -125,19 +140,22 @@ export class TaskController {
     return category;
   };
 
-  removeTaskById = index => {
-    this.model.localData[0].forEach(column => {
-      for (let i = 0; i < column.items.length; i++) {
-        if (column.items[i].index === index) {
-          column.items.splice(i, 1);
-          break;
+  removeTaskById = (index: number) => {
+    if (this.model.localData) {
+      this.model.localData[0]?.forEach(column => {
+        for (let i = 0; i < column.items.length; i++) {
+          if (column.items[i].index === index) {
+            column.items.splice(i, 1);
+            break;
+          }
         }
-      }
-    });
+      });
+    }
   };
 
-  addRemoveListener = task => {
-    const icon = document.getElementById(task.index)!.children[0].children[1];
+  addRemoveListener = (task: TaskVO) => {
+    const icon = document.getElementById(task.index.toString())!.children[0]
+      .children[1];
     icon.addEventListener('click', e => {
       this.processRemoval(e);
     });
@@ -153,22 +171,30 @@ export class TaskController {
   };
 
   // this function was converted to arrow function while doing TS
-  processRemoval = e => {
+  processRemoval = (e: Event) => {
+    // to do: think about better solution
     const task = document.getElementById(
-      e.target.parentElement.parentElement.id
+      (e.target as HTMLElement).parentElement!.parentElement!.id
     );
     if (task) {
       task.remove();
       let taskDB;
-      this.model.localData[0].forEach(column => {
-        column.items.forEach(item => {
-          if (item.index === parseInt(task.id)) {
-            taskDB = item;
-          }
+      if (this.model.localData) {
+        this.model.localData[0]?.forEach((column: ListVO) => {
+          column.items.forEach(item => {
+            if (item.index === parseInt(task.id)) {
+              taskDB = item;
+            }
+          });
         });
-      });
-      this.removeTaskById(parseInt(e.target.parentElement.parentElement.id));
-      this.model.taskService.delete(taskDB);
+      }
+      // here as well: better solution needed
+      this.removeTaskById(
+        parseInt((e.target as HTMLElement).parentElement!.parentElement!.id)
+      );
+      if (taskDB) {
+        this.model.taskService.delete(taskDB);
+      }
     }
   };
 
@@ -182,27 +208,29 @@ export class TaskController {
   };
 
   // this function was converted to arrow function while doing TS
-  processSubmit = async e => {
+  processSubmit = async (e: Event) => {
     e.preventDefault();
     const body = <HTMLTextAreaElement>document.querySelector('#new-card-text');
     if (body.value !== '') {
       const task = this.createTask();
       const taskView = new TaskView(task);
       const column = document.querySelectorAll('.col-body')[1];
-      taskView.render(column, 'beforeend');
+      taskView.render(column as HTMLElement, 'beforeend');
       this.colorTask(task);
       this.addRemoveListener(task);
       this.addOnDragStartListener(task);
       this.addTaskEditListener(task);
       this.model.taskService.add(task);
-      this.model.counterService.update(this.model.localData[2]);
+      if (this.model.localData) {
+        this.model.counterService.update(this.model.localData[2]);
+      }
       this.model.localData = await this.model.dataService.get();
       this.setTasks();
     }
   };
 
-  addOnDragStartListener = task => {
-    const item = document.getElementById(task.index);
+  addOnDragStartListener = (task: TaskVO) => {
+    const item = document.getElementById(task.index.toString());
     if (item) {
       item.addEventListener('dragstart', e => {
         e.dataTransfer!.setData('text/plain', (e.target! as HTMLElement).id);
@@ -230,25 +258,35 @@ export class TaskController {
           (dropzone as HTMLElement).appendChild(draggableElement!);
           (e as DragEvent).dataTransfer!.clearData();
 
-          let columnFrom;
-          this.model.localData[0].forEach(column => {
-            column.items.forEach(task => {
-              if (task.index === parseInt(draggableElement!.id)) {
-                columnFrom = column;
-              }
+          let columnFrom: ListVO = {
+            label: '',
+            items: [],
+            _id: '',
+          };
+          if (this.model.localData) {
+            this.model.localData[0].forEach((column: ListVO) => {
+              column.items.forEach((task: TaskVO) => {
+                if (task.index === parseInt(draggableElement!.id)) {
+                  columnFrom = column;
+                }
+              });
             });
-          });
+          }
 
-          let columnTo;
+          let columnTo: ListVO = {
+            label: '',
+            items: [],
+            _id: '',
+          };
           const columns = document.querySelectorAll('.col-body');
           for (let i = 1; i < columns.length; i++) {
-            if (columns[i] === dropzone) {
+            if (columns[i] === dropzone && this.model.localData) {
               columnTo = this.model.localData[0][i - 1];
               break;
             }
           }
 
-          let taskIndex;
+          let taskIndex: number = 0;
           for (let i = 0; i < columnFrom.items.length; i++) {
             if (columnFrom.items[i].index === parseInt(draggableElement!.id)) {
               taskIndex = i;
@@ -269,44 +307,50 @@ export class TaskController {
     const pickers = document.querySelectorAll('*[id^="c-"]');
     pickers.forEach(picker => {
       picker.addEventListener('change', e => {
-        this.model.localData[1].forEach(category => {
-          if (category.index === picker.id) {
-            category.color = (e.target! as HTMLInputElement).value;
-            console.log(category);
-            this.model.categoryService.edit(category);
-          }
-        });
-        this.model.localData[0].forEach(column => {
-          column.items.forEach(item => {
-            if (item.category === picker.id) {
-              document.getElementById(item.index)!.style.backgroundColor = (
-                e.target! as HTMLInputElement
-              ).value;
+        if (this.model.localData) {
+          this.model.localData[1].forEach((category: CategoryVO) => {
+            if (category.index === picker.id) {
+              category.color = (e.target as HTMLInputElement).value;
+              this.model.categoryService.edit(category);
             }
           });
-        });
-        console.log(this.model.localData[1]);
+        }
+        if (this.model.localData) {
+          this.model.localData[0].forEach((column: ListVO) => {
+            column.items.forEach(item => {
+              if (item.category === picker.id) {
+                document.getElementById(
+                  item.index.toString()
+                )!.style.backgroundColor = (
+                  e.target! as HTMLInputElement
+                ).value;
+              }
+            });
+          });
+        }
       });
     });
   };
 
-  addTaskEditListener = task => {
-    const contentEditable = document.getElementById(task.index)!.children[1]
-      .children[0];
+  addTaskEditListener = (task: TaskVO) => {
+    const contentEditable = document.getElementById(task.index.toString())!
+      .children[1].children[0];
     contentEditable.addEventListener('blur', () => {
-      this.model.localData[0].forEach(column => {
-        column.items.forEach(item => {
-          if (
-            item.index ===
-            parseInt(
-              (contentEditable.parentNode!.parentNode! as HTMLElement).id
-            )
-          ) {
-            item.body = (contentEditable as HTMLElement).innerText;
-          }
+      if (this.model.localData) {
+        this.model.localData[0].forEach((column: ListVO) => {
+          column.items.forEach(item => {
+            if (
+              item.index ===
+              parseInt(
+                (contentEditable.parentNode!.parentNode! as HTMLElement).id
+              )
+            ) {
+              item.body = (contentEditable as HTMLElement).innerText;
+            }
+          });
         });
-      });
-      this.model.taskService.edit(task);
+        this.model.taskService.edit(task);
+      }
     });
   };
 
@@ -316,12 +360,14 @@ export class TaskController {
     this.addOnDropListener();
     this.addColorChangeListeners();
 
-    this.model.localData[0].forEach(column => {
-      column.items.forEach(task => {
-        this.addRemoveListener(task);
-        this.addTaskEditListener(task);
-        this.addOnDragStartListener(task);
+    if (this.model.localData) {
+      this.model.localData[0].forEach((column: ListVO) => {
+        column.items.forEach(task => {
+          this.addRemoveListener(task);
+          this.addTaskEditListener(task);
+          this.addOnDragStartListener(task);
+        });
       });
-    });
+    }
   };
 }
